solhint issues resolved

+++ contracts/serverless-v2/RelaySubscriptions.sol
@@ -1,16 +1,15 @@
 // SPDX-License-Identifier: UNLICENSED
 pragma solidity ^0.8.0;
 
-import "@openzeppelin/contracts-upgradeable/proxy/utils/Initializable.sol";
-import "@openzeppelin/contracts-upgradeable/utils/ContextUpgradeable.sol";
-import "@openzeppelin/contracts-upgradeable/utils/introspection/ERC165Upgradeable.sol";
-import "@openzeppelin/contracts-upgradeable/access/AccessControlUpgradeable.sol";
-import "@openzeppelin/contracts-upgradeable/proxy/utils/UUPSUpgradeable.sol";
-import "@openzeppelin/contracts/token/ERC20/IERC20.sol";
-import "@openzeppelin/contracts/token/ERC20/utils/SafeERC20.sol";
-import "@openzeppelin/contracts/utils/cryptography/ECDSA.sol";
-import "../interfaces/IAttestationVerifier.sol";
-import "./Relay.sol";
+import {Initializable} from "@openzeppelin/contracts-upgradeable/proxy/utils/Initializable.sol";
+import {ContextUpgradeable} from "@openzeppelin/contracts-upgradeable/utils/ContextUpgradeable.sol";
+import {ERC165Upgradeable} from "@openzeppelin/contracts-upgradeable/utils/introspection/ERC165Upgradeable.sol";
+import {AccessControlUpgradeable} from "@openzeppelin/contracts-upgradeable/access/AccessControlUpgradeable.sol";
+import {UUPSUpgradeable} from "@openzeppelin/contracts-upgradeable/proxy/utils/UUPSUpgradeable.sol";
+import {IERC20} from "@openzeppelin/contracts/token/ERC20/IERC20.sol";
+import {SafeERC20} from "@openzeppelin/contracts/token/ERC20/utils/SafeERC20.sol";
+import {ECDSA} from "@openzeppelin/contracts/utils/cryptography/ECDSA.sol";
+import {Relay} from "./Relay.sol";
 
 /**
  * @title RelaySubscriptions Contract
@@ -61,6 +60,7 @@ contract RelaySubscriptions is
     }
 
     /// @inheritdoc UUPSUpgradeable
+    // solhint-disable-next-line no-empty-blocks
     function _authorizeUpgrade(address /*account*/) internal view override onlyRole(DEFAULT_ADMIN_ROLE) {}
 
     //-------------------------------- Overrides end --------------------------------//
@@ -522,9 +522,10 @@ contract RelaySubscriptions is
         Relay.Job memory _job,
         bytes calldata _output,
         uint8 _errorCode
-    ) internal returns (bool success, uint callbackGas) {
+    ) internal returns (bool success, uint256 callbackGas) {
         if (tx.gasprice <= _job.maxGasPrice) {
-            uint startGas = gasleft();
+            uint256 startGas = gasleft();
+            // solhint-disable-next-line avoid-low-level-calls
             (success, ) = _job.callbackContract.call{gas: _job.callbackGasLimit}(
                 abi.encodeWithSignature(
                     "oysterResultCall(uint256,address,bytes32,bytes,bytes,uint8)",
@@ -549,6 +550,7 @@ contract RelaySubscriptions is
         // TODO: If paySuccess is false then deposit will be stucked forever. Find a way out.
         // transfer callback cost to gateway
         _callbackCost = _callbackCost > _callbackDeposit ? _callbackDeposit : _callbackCost;
+        // solhint-disable-next-line no-unused-vars, avoid-low-level-calls
         (bool paySuccess, ) = _gatewayOwner.call{value: _callbackCost}("");
 
         // transfer remaining native asset to the jobOwner
@@ -633,6 +635,7 @@ contract RelaySubscriptions is
 
         RELAY.TOKEN().safeTransfer(_jobOwner, usdcAmount);
         // TODO: do we need to check this bool success
+        // solhint-disable-next-line avoid-low-level-calls
         (bool success, ) = _jobOwner.call{value: callbackAmount}("");
 
         emit JobSubscriptionDepositsRefunded(_jobSubsId, _jobOwner, usdcAmount, callbackAmount, success);
